Задача 3

+++ src/Main.java
@@ -4,8 +4,8 @@ public class Main {
     public static void main(String[] args) {
         System.out.println("Домашка 9, МАССИВЫ, часть 2, от 09.11.2022 - Задача 3"); // о, ну это было легко!
         System.out.println("Траты по дням за месяц составили " + Arrays.toString(arr));
-        int summExpensesMonth = 0;
-        int middleExpensesMonth = 0;
+        double summExpensesMonth = 0;
+        double middleExpensesMonth = 0;
         for (int i = 0; i < arr.length; i++) {
             summExpensesMonth += arr[i];
             middleExpensesMonth = summExpensesMonth / arr.length;
